+++ SungJinwooRobot/modules/sql/antiflood_sql.py
@@ -1,5 +1,5 @@
 import threading
-
+from sqlalchemy.sql.sqltypes import BigInteger
 from sqlalchemy import String, Column, Integer, UnicodeText
 
 from SungJinwooRobot.modules.sql import SESSION, BASE
@@ -11,7 +11,7 @@ DEF_OBJ = (None, DEF_COUNT, DEF_LIMIT)
 class FloodControl(BASE):
     __tablename__ = "antiflood"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     count = Column(Integer, default=DEF_COUNT)
     limit = Column(Integer, default=DEF_LIMIT)
