+++ SungJinwooRobot/modules/sql/userinfo_sql.py
@@ -2,11 +2,11 @@ import threading
 
 from SungJinwooRobot.modules.sql import BASE, SESSION
 from sqlalchemy import Column, Integer, UnicodeText
-
+from sqlalchemy.sql.sqltypes import BigInteger
 
 class UserInfo(BASE):
     __tablename__ = "userinfo"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     info = Column(UnicodeText)
 
     def __init__(self, user_id, info):
